+++ src/lib.rs
@@ -106,10 +106,6 @@
 //!
 //! * When starting Actix actors from outside Tokio/Actix, [`woab::block_on`](block_on) must be
 //!   used. This is a limitation of Actix that needs to be respected.
-//! * Some GTK actions (like removing a widget) can fire signals synchronously. If these signals
-//!   are registered as builder signals, WoAB will not be able to route them and panic because it
-//!   will happen while the Actix runtime is occupied. To work around this, use
-//!   [`woab::spawn_outside`](spawn_outside).
 //! * `dialog.run()` must not be used - use [`woab::run_dialog`](crate::run_dialog) instead.
 //! * If an actor is created inside a `gtk::Application::connect_activate`, its `started` method
 //!   will run **after** the `activate` signal is done. This can be a problem for methods like